Move ExtractTraceID from util to util/tracing
