+++ pkg/util/tracing/tracing.go
@@ -1,4 +1,4 @@
-package util
+package tracing
 
 import (
 	"context"
